requirements.txt
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
openai>=1.0.0
faiss-cpu>=1.7.0
voyageai>=0.2.0
python-dotenv>=1.0.0
numpy>=1.24.0
